@@ -1,23 +1,14 @@
-// import "@babel/polyfill";
-import "mutationobserver-shim";
-import Vue from "vue";
-import "./plugins/bootstrap-vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
-import store from "./store";
-// const VCalendar = require('v-calendar');
-import VCalendar from 'v-calendar';
+// import store from "./store";
+import "mutationobserver-shim";
 import '@fortawesome/fontawesome-free/js/all'
 
-// Use v-calendar & v-date-picker components
-Vue.use(VCalendar, {
-  componentPrefix: 'vc',  // Use <vc-calendar /> instead of <v-calendar />
-});
 
-Vue.config.productionTip = false;
+import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap";
+// import VCalendar from 'v-calendar';
 
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount("#app");
+
+createApp(App).use(router).mount('#app');
